Fix color conversion in color advice
The colors in the maps of the color advice did not match up
exactly with other colors. For example, the viridis colors
were not exactly the same as those in matplotlib, where the
color map originated.

The problem turned out to be with the color conversions in
the Python colormap library. Specifically, the illumination
got changed. The default illumination used when converting
from sRGB space to LAB is different than the default illumination
you get when directly creating a LabColor object. That means
that conversion to LAB happened with one illumination and
conversion back to sRGB happened in a different illumination.

Fix the scripts and resulting tables to have explicit and
consistent illumination.